+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/_help.py
@@ -459,12 +459,15 @@ examples:
   - name: Create a Linux task using a private GitHub repository which builds the hello-world image without triggers on Arm architecture (V7 variant)
     text: >
         az acr task create -t hello-world:{{.Run.ID}} -n hello-world -r MyRegistry -c https://github.com/Azure-Samples/acr-build-helloworld-node.git -f Dockerfile --commit-trigger-enabled false --pull-request-trigger-enabled false --git-access-token 0000000000000000000000000000000000000000 --platform linux/arm/v7
-  - name: Create a Linux task from a public GitHub repository which builds the hello-world image with a git commit trigger
+  - name: Create a Linux task from a public GitHub repository which builds the hello-world image with a git commit trigger. Note that this task does not use Source Registry (MyRegistry), so we can explicitly set Auth mode as None for it.
     text: >
-        az acr task create -t hello-world:{{.Run.ID}} -n hello-world -r MyRegistry -c https://github.com/Azure-Samples/acr-build-helloworld-node.git -f Dockerfile --git-access-token 0000000000000000000000000000000000000000
+        az acr task create -t hello-world:{{.Run.ID}} -n hello-world -r MyRegistry --auth-mode None -c https://github.com/Azure-Samples/acr-build-helloworld-node.git -f Dockerfile --git-access-token 0000000000000000000000000000000000000000
   - name: Create a Windows task from a public GitHub repository which builds the Azure Container Builder image on Amd64 architecture.
     text: >
         az acr task create -t acb:{{.Run.ID}} -n acb-win -r MyRegistry -c https://github.com/Azure/acr-builder.git -f Windows.Dockerfile --commit-trigger-enabled false --pull-request-trigger-enabled false --platform Windows/amd64
+  - name: Create a task from a public GitHub repository which builds and pushes the image in 2 registries. One is the MyRegistry1 and the other one in this case is MyRegistry2 (per testtask.yaml file). The source registry credentials (MyRegistry1) are auto generated, while you can pass in credentials for MyRegistry2 using 'login-server', 'username' and 'password' flags
+    text: >
+      az acr task create -n hello-world -r MyRegistry1 -c https://github.com/Azure-Samples/acr-tasks.git#:multipleRegistries -f testtask.yaml --values values.yaml --login-server 'MyRegistry2.azurecr.io' --username 'registryUsername' --password 'registrySecret' --commit-trigger-enabled false --pull-request-trigger-enabled false
 """
 
 helps['acr task delete'] = """
@@ -562,6 +565,9 @@ examples:
   - name: Update platform for the Build step of your Task to Windows (prev Linux).
     text: >
         az acr task update -n MyTask -r MyRegistry --platform Windows
+  - name: To remove a credential from a task, you can update the task by sending empty username and password for the Registry. For instance, Remove credentials for MyRegistry2.
+    text: >
+      az acr task update -n hello-world -r MyRegistry --credential 'MyRegistry1.azurecr.io' --username '' --password ''
 """
 
 helps['acr task update-run'] = """
@@ -571,6 +577,47 @@ examples:
   - name: Update an existing run to be archived.
     text: >
         az acr task update-run -r MyRegistry --run-id runId --no-archive false
+"""
+
+helps['acr task credential'] = """
+type: group
+short-summary: Credential for a task
+"""
+
+helps['acr task credential add'] = """
+type: group
+short-summary: Add a credential to the task
+examples:
+  - name: Enable the system-assigned credential to an existing task. This will replace all existing user-assigned identities.
+    text: >
+        az acr task credential add -n taskname -r registryname --login-server myregistry.docker.io --username 'myusername' --password 'mysecret'
+"""
+
+helps['acr task credential remove'] = """
+type: group
+short-summary: Remove credential for a task.
+examples:
+  - name: Remove a user-assigned credential from a task.
+    text: >
+        az acr task credential remove -n taskname -r registryname --login-server myregistry.docker.io
+"""
+
+helps['acr task credential update'] = """
+type: group
+short-summary: Update the credential for a task.
+examples:
+  - name: Update the credential for a task
+    text: >
+        az acr task credential update -n taskname -r registryname --login-server myregistry.docker.io --username 'myusername2' --password 'mysecret'
+"""
+
+helps['acr task credential show'] = """
+type: group
+short-summary: Display the Credentials for task.
+examples:
+  - name: Display the Credentials for a task.
+    text: >
+        az acr task credential show -n taskname -r registryname
 """
 
 helps['acr update'] = """