+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/build.py
@@ -13,7 +13,7 @@ from knack.log import get_logger
 from knack.util import CLIError
 from azure.cli.core.commands import LongRunningOperation
 
-from ._utils import validate_managed_registry, get_validate_platform
+from ._utils import validate_managed_registry, get_validate_platform, get_custom_registry_credentials
 from ._stream_utils import stream_logs
 from ._archive_utils import upload_source_code, check_remote_source_code
 
@@ -38,7 +38,8 @@ def acr_build(cmd,  # pylint: disable=too-many-locals
               no_logs=False,
               os_type=None,
               platform=None,
-              target=None):
+              target=None,
+              auth_mode=None):
     _, resource_group_name = validate_managed_registry(
         cmd, registry_name, resource_group_name, BUILD_NOT_SUPPORTED)
 
@@ -116,7 +117,11 @@ def acr_build(cmd,  # pylint: disable=too-many-locals
         docker_file_path=docker_file_path,
         timeout=timeout,
         arguments=(arg if arg else []) + (secret_arg if secret_arg else []),
-        target=target
+        target=target,
+        credentials=get_custom_registry_credentials(
+            cmd,
+            auth_mode=auth_mode
+        )
     )
 
     queued_build = LongRunningOperation(cmd.cli_ctx)(client_registries.schedule_run(