+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/commands.py
@@ -174,6 +174,10 @@ def load_command_table(self, _):  # pylint: disable=too-many-statements
         g.command('list', 'acr_task_list')
         g.command('delete', 'acr_task_delete')
         g.command('update', 'acr_task_update')
+        g.command('credential add', 'acr_task_credential_add')
+        g.command('credential remove', 'acr_task_credential_remove')
+        g.command('credential update', 'acr_task_credential_update')
+        g.command('credential show', 'acr_task_credential_show')
         g.command('run', 'acr_task_run', client_factory=cf_acr_runs,
                   table_transformer=run_output_format, supports_no_wait=True)
         g.command('list-runs', 'acr_task_list_runs', client_factory=cf_acr_runs,